requirements.txt
ultralytics>=8.0.0
torch>=1.7.0
torchvision>=0.8.1
opencv-python>=4.5.0
pillow>=7.1.2
pyyaml>=5.3.1
tqdm>=4.41.0
matplotlib>=3.2.2
seaborn>=0.11.0
pandas>=1.1.4
numpy>=1.18.5
requests>=2.23.0
PyYAML>=5.1
